First Round
I attempted to tackle many javascript/jquery problems I anticipated.
The zen rock garden definitely needs to be refined but I was able to
create a div in which you can draw in. The fading elements were one of
the most difficult parts to figure out, and I still need to figure out
how to make them fade out.